[BE]hotfix: User 관계 있는 오류 수정

See merge request s11-webmobile1-sub2/S11P12B301!58

+++ BE/src/main/java/com/example/backend/reservation/service/ReserveService.java
@@ -41,10 +41,10 @@ public class ReserveService {
     // 상담 예약 추가
     @Transactional
     public void reserve(ReserveRequestDto request) {
-        User user = userRepository.findById((int) request.getUserIdx())
+        User user = userRepository.findById(request.getUserIdx())
                 .orElseThrow(() -> new RuntimeException("User not ,found"));
 
-        User counselor = userRepository.findById((int) request.getCoIdx())
+        User counselor = userRepository.findById(request.getCoIdx())
                 .orElseThrow(() -> new RuntimeException("Counselor not found"));
 
         LocalDate reservationDate = request.getReserveInfoDate();
